Refactor Patrolling Enemy
