+++ Assets/PatrolType.cs
@@ -1,0 +1,42 @@
+﻿using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class PatrolType : MonoBehaviour
+{
+    private bool movingRight;
+    private Transform groundDetection;
+    public float speed;
+
+    void Start()
+    {
+        groundDetection = transform.GetChild(0);
+        speed = 3.5f;
+    }
+
+    void Update()
+    {
+        Patrol();   
+    }
+
+    private void Patrol()
+    {
+        transform.Translate(Vector2.right * speed * Time.deltaTime);
+
+        RaycastHit2D groundInfo = Physics2D.Raycast(groundDetection.position, Vector2.down, 2f);
+
+        if (groundInfo.collider == false)
+        {
+            if (movingRight == true)
+            {
+                transform.eulerAngles = new Vector3(0, -180, 0);
+                movingRight = false;
+            }
+            else
+            {
+                transform.eulerAngles = new Vector3(0, 0, 0);
+                movingRight = true;
+            }
+        }
+    }
+}
