+++ Assets/PatrolType.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c85b1782434cd7e4293234c9d5bcf558
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
